+++ lib/get_state.dart
@@ -2,3 +2,5 @@ library getstate;
 
 export 'package:get_state/src/abs.dart';
 export 'package:get_state/src/recorder.dart';
+
+export 'package:freezed_annotation/freezed_annotation.dart';